requirements.txt
pyfunctional
pykakasi